removed more paywall components
